requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt pyproject.toml
accelerate==1.11.0
    # via
    #   266-final-project (pyproject.toml)
    #   transformers
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via fsspec
aiosignal==1.4.0
    # via aiohttp
anyio==4.11.0
    # via httpx
anytree==2.13.0
    # via openhownet
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonlines
beautifulsoup4==4.14.2
    # via gdown
bert-score==0.3.13
    # via textattack
bioc==2.1
    # via flair
boto3==1.41.5
    # via flair
botocore==1.41.5
    # via
    #   boto3
    #   s3transfer
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via nltk
conllu==4.5.3
    # via flair
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==4.4.1
    # via
    #   266-final-project (pyproject.toml)
    #   textattack
deprecated==1.3.1
    # via flair
dill==0.4.0
    # via
    #   datasets
    #   multiprocess
    #   pyfunctional
docopt==0.6.2
    # via
    #   bioc
    #   num2words
editdistance==0.8.1
    # via textattack
filelock==3.20.0
    # via
    #   datasets
    #   gdown
    #   huggingface-hub
    #   textattack
    #   torch
    #   transformers
flair==0.15.1
    # via textattack
fonttools==4.60.1
    # via matplotlib
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
ftfy==6.3.1
    # via flair
gdown==5.2.0
    # via flair
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via datasets
huggingface-hub==0.36.0
    # via
    #   266-final-project (pyproject.toml)
    #   accelerate
    #   datasets
    #   flair
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
intervaltree==3.1.0
    # via bioc
jieba==0.42.1
    # via textattack
jinja2==3.1.6
    # via
    #   mpld3
    #   torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.2
    # via
    #   nltk
    #   scikit-learn
jsonlines==4.0.0
    # via bioc
kiwisolver==1.4.9
    # via matplotlib
langdetect==1.0.9
    # via flair
language-tool-python==3.1.0
    # via textattack
lemminflect==0.2.3
    # via textattack
lru-dict==1.4.1
    # via textattack
lxml==6.0.2
    # via
    #   bioc
    #   flair
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.7
    # via
    #   bert-score
    #   flair
    #   mpld3
    #   seaborn
more-itertools==10.8.0
    # via
    #   flair
    #   textattack
mpld3==0.5.12
    # via flair
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.18
    # via datasets
networkx==3.5
    # via torch
nltk==3.9.2
    # via textattack
num2words==0.5.14
    # via textattack
numpy==2.3.4
    # via
    #   accelerate
    #   bert-score
    #   contourpy
    #   datasets
    #   lemminflect
    #   matplotlib
    #   pandas
    #   patsy
    #   pytorch-revgrad
    #   scikit-learn
    #   scipy
    #   seaborn
    #   sklearn-utils
    #   statsmodels
    #   textattack
    #   transformers
openhownet==2.0
    # via textattack
packaging==25.0
    # via
    #   accelerate
    #   bert-score
    #   datasets
    #   huggingface-hub
    #   language-tool-python
    #   matplotlib
    #   statsmodels
    #   transformers
pandas==2.3.3
    # via
    #   bert-score
    #   datasets
    #   seaborn
    #   sklearn-utils
    #   statsmodels
    #   textattack
patsy==1.0.2
    # via statsmodels
pillow==12.0.0
    # via matplotlib
pinyin==0.4.0
    # via textattack
pptree==3.1
    # via flair
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.1
    # via transformers
psutil==7.1.3
    # via
    #   accelerate
    #   language-tool-python
pyarrow==22.0.0
    # via datasets
pyfunctional==1.5.0
    # via sklearn-utils
pyparsing==3.2.5
    # via matplotlib
pysocks==1.7.1
    # via requests
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   flair
    #   matplotlib
    #   pandas
pytorch-revgrad==0.2.0
    # via flair
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   transformers
regex==2025.11.3
    # via
    #   flair
    #   nltk
    #   segtok
    #   transformers
requests==2.32.5
    # via
    #   bert-score
    #   datasets
    #   gdown
    #   huggingface-hub
    #   language-tool-python
    #   openhownet
    #   transformers
    #   wikipedia-api
s3transfer==0.15.0
    # via boto3
safetensors==0.6.2
    # via
    #   accelerate
    #   transformers
scikit-learn==1.7.2
    # via
    #   flair
    #   sklearn-utils
scipy==1.16.3
    # via
    #   266-final-project (pyproject.toml)
    #   scikit-learn
    #   sklearn-utils
    #   statsmodels
    #   textattack
seaborn==0.13.2
    # via sklearn-utils
segtok==1.5.11
    # via flair
sentencepiece==0.2.1
    # via transformers
setuptools==80.9.0
    # via
    #   openhownet
    #   torch
six==1.17.0
    # via
    #   langdetect
    #   python-dateutil
sklearn-utils==0.0.15
    # via 266-final-project (pyproject.toml)
sniffio==1.3.1
    # via anyio
sortedcontainers==2.4.0
    # via intervaltree
soupsieve==2.8
    # via beautifulsoup4
sqlitedict==2.1.0
    # via flair
statsmodels==0.14.5
    # via sklearn-utils
sympy==1.14.0
    # via torch
tabulate==0.9.0
    # via
    #   flair
    #   pyfunctional
terminaltables==3.1.10
    # via textattack
textattack==0.3.10
    # via 266-final-project (pyproject.toml)
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.1
    # via transformers
toml==0.10.2
    # via language-tool-python
torch==2.9.0
    # via
    #   266-final-project (pyproject.toml)
    #   accelerate
    #   bert-score
    #   flair
    #   pytorch-revgrad
    #   textattack
    #   transformer-smaller-training-vocab
    #   transformers
tqdm==4.67.1
    # via
    #   bert-score
    #   bioc
    #   datasets
    #   flair
    #   gdown
    #   huggingface-hub
    #   language-tool-python
    #   nltk
    #   openhownet
    #   textattack
    #   transformers
transformer-smaller-training-vocab==0.4.2
    # via flair
transformers==4.57.1
    # via
    #   266-final-project (pyproject.toml)
    #   bert-score
    #   flair
    #   textattack
    #   transformer-smaller-training-vocab
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   huggingface-hub
    #   torch
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   botocore
    #   requests
wcwidth==0.2.14
    # via ftfy
wikipedia-api==0.8.1
    # via flair
word2number==1.1
    # via textattack
wrapt==2.0.1
    # via deprecated
xxhash==3.6.0
    # via datasets
yarl==1.22.0
    # via aiohttp
